[#61] :sparkles: refactor: Input 컴포넌트 number 타입 분리
