+++ client/src/components/Input/index.tsx
@@ -1,1 +1,4 @@
-export { default } from './Input';
+import Input from './Input';
+import NumberInput from './NumberInput';
+
+export { Input, NumberInput };
